Fix issues found
